feat: [IDP-143]: refactor
