+++ idp-service/modules/events/src/main/java/io/harness/SecretCRUDListener.java
@@ -5,7 +5,7 @@
  * https://polyformproject.org/wp-content/uploads/2020/05/PolyForm-Free-Trial-1.0.0.txt.
  */
 
-package io.harness.event;
+package io.harness.events.src.main.java.io.harness;
 
 import io.harness.eventsframework.consumer.Message;
 import io.harness.ng.core.event.MessageListener;
