+++ idp-service/service/BUILD.bazel
@@ -37,7 +37,7 @@ java_binary(
         "//idp-service/contracts:module",
         "//idp-service/service/src/main/java/io/harness/idp/config:module",
         "//idp-service/service/src/main/java/io/harness/idp/secret:module",
-        "//idp-service/modules/utils/k8s:module",
+        "//idp-service/modules/k8s:module",
         "@maven//:com_google_guava_guava",
         "@maven//:com_github_dirkraft_dropwizard_file_assets_dropwizard_file_assets",
         "@maven//:com_google_inject_guice",
